Removes unneeded sudo, adds clean slate in demo
The sudo to recursively copy the target files from bob's machine
to carl's machine is actually not needed.
This commit removes it from the demo/README.md

The commit also adds a section to the README that explains how
to use the run_demo script to clean all files delated during the
demo which is necessary to re-run the demo.