Pin in-toto version and update instructions
in-toto is still in a pre-release phase and might change
in ways that are not compatible with this demo.

For now we pin the demo to a compatible commit id with a newly
added requirements file and update the installation instructions
to install in-toto using that file.

As soon as we have a versioned release we should change the sha
to the version number.