Add script to run demo code in README
Add a script that extracts the demo code snippets from README.md
and runs them in a shell, raising `SystemExit`, if the output is
not as expected.

This is an automated testing alternative to the existing
`run_demo.py`, which replicates the commands from the demo
instructions, with the advantage that the commands don't have to be
synced.

NOTE: The script requires the in-toto version specified in
requirements.txt, i.e. 0.2.3 at the moment.