Filter deprecation warnings on EOL-ed Pythons
The demo tests compare expected output to actual output of the
demo run, which since recently shows Deprecation warnings on EOLed
Python 2.7 and 3.5 tests.

As a quickfix we just strip those warnings from the demo run
output. If we should continue support for 2.7 and 3.5 needs to be
decided in in-toto not her in the demo.