Fix test by updating expected output
PRs #21 and #22 made some changes in the demo instructions
(README.md) but didn't adopt them in the expected output of the
demo test script (run_demo_md.py), thus breaking the test.

This likely wans't detected when merging the PRs because CI too
has stopped working a while ago.

Besides updating the expected output in the test script, this
commit:
- removes some invisible characters from README.md
- removes a stray mention of `virtualenvwrapper` (left over from
  before #21)
- makes the sed command added in #21 BSD compatible

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>